Adds Fielding stats endpoint and tests it

+++ PowerUp/PowerUp/Fetchers/MLBLookupService/HittingStatsResult.cs
@@ -22,7 +22,7 @@ namespace PowerUp.Fetchers.MLBLookupService
   {
     public int LSPlayerId { get; }
     public int Year { get; }
-    public double TeamSeq { get; }
+    public int TeamSeq { get; }
     
     public int GamesPlayed { get; }
     public int AtBats { get; }
@@ -63,7 +63,7 @@ namespace PowerUp.Fetchers.MLBLookupService
     {
       LSPlayerId = int.Parse(result.player_id!);
       Year = int.Parse(result.season!);
-      TeamSeq = double.Parse(result.team_seq!);
+      TeamSeq = (int)double.Parse(result.team_seq!);
       GamesPlayed = int.Parse(result.g!);
       AtBats = int.Parse(result.ab!);
       PlateAppearances = int.Parse(result.tpa!);
